requirements.txt
elasticsearch==8.12.0
groq>=0.4.0
python-dotenv==1.0.0
gradio>=4.44.0
